# docs/api/openapi/api/openapi.yaml
openapi: 3.1.0
info:
  title: Chronos API
  description: "Shift, leave, and sickness planning \u2014 FS Phase 1 implementation."
  version: 0.2.0
paths:
...
  /api/users:
    get:
      tags:
      - users
      summary: List Users
      operationId: list_users_api_users_get
      parameters:
      - name: team_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Team Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDetail'
                title: Response List Users Api Users Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - users
      summary: Create User
      operationId: create_user_api_users_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/users/{user_id}:
    get:
      tags:
      - users
      summary: Get User
      operationId: get_user_api_users__user_id__get
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - users
      summary: Update User
      operationId: update_user_api_users__user_id__patch
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - users
      summary: Deactivate
      operationId: deactivate_api_users__user_id__delete
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/users/{user_id}/role:
    put:
      tags:
      - users
      summary: Set Role
      operationId: set_role_api_users__user_id__role_put
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleChange'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/users/me/password:
    post:
      tags:
      - users
      summary: Change My Password
      description: 'H-04: self-service rotation.


        Verifies the current password, enforces the policy on the new one, and

        stamps ``password_changed_at`` + clears ``must_rotate_password`` so the

        forced-rotate banner goes away. Emits an audit event in the same

        transaction.'
      operationId: change_my_password_api_users_me_password_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SelfPasswordChange'
        required: true
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/users/{user_id}/password:
    put:
      tags:
      - users
      summary: Set Password
      operationId: set_password_api_users__user_id__password_put
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordUpdate'
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/workflows:
    get:
      tags:
      - workflows
      summary: List Workflows
      operationId: list_workflows_api_workflows_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/WorkflowOut'
                type: array
                title: Response List Workflows Api Workflows Get
    post:
      tags:
      - workflows
      summary: Create Workflow
      operationId: create_workflow_api_workflows_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/workflows/{workflow_id}:
    get:
      tags:
      - workflows
      summary: Get Workflow
      operationId: get_workflow_api_workflows__workflow_id__get
      parameters:
      - name: workflow_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Workflow Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - workflows
      summary: Update Workflow
      operationId: update_workflow_api_workflows__workflow_id__put
      parameters:
      - name: workflow_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Workflow Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - workflows
      summary: Delete Workflow
      operationId: delete_workflow_api_workflows__workflow_id__delete
      parameters:
      - name: workflow_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Workflow Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PasswordUpdate:
      properties:
        new_password:
          type: string
          minLength: 8
          title: New Password
      type: object
      required:
      - new_password
      title: PasswordUpdate
...
    Role:
      type: string
      enum:
      - employee
      - team_lead
      - hr
      - admin
      title: Role
    RoleChange:
      properties:
        role:
          $ref: '#/components/schemas/Role'
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
      type: object
      required:
      - role
      title: RoleChange
    SelfPasswordChange:
      properties:
        current_password:
          type: string
          minLength: 1
          title: Current Password
        new_password:
          type: string
          minLength: 8
          title: New Password
      type: object
      required:
      - current_password
      - new_password
      title: SelfPasswordChange
...
    UserCreate:
      properties:
        email:
          type: string
          title: Email
        password:
          type: string
          minLength: 8
          title: Password
        first_name:
          anyOf:
          - type: string
            maxLength: 120
          - type: 'null'
          title: First Name
        last_name:
          anyOf:
          - type: string
            maxLength: 120
          - type: 'null'
          title: Last Name
        role:
          $ref: '#/components/schemas/Role'
          default: employee
        locale:
          type: string
          title: Locale
          default: en
        time_zone:
          type: string
          title: Time Zone
          default: UTC
        department_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Department Id
        team_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Team Id
      type: object
      required:
      - email
      - password
      title: UserCreate
    UserDetail:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        email:
          type: string
          title: Email
        first_name:
          anyOf:
          - type: string
          - type: 'null'
          title: First Name
        last_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Name
        role:
          type: string
          title: Role
        locale:
          type: string
          title: Locale
        time_zone:
          type: string
          title: Time Zone
        department_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Department Id
        team_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Team Id
        must_rotate_password:
          type: boolean
          title: Must Rotate Password
          default: false
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - email
      - role
      - locale
      - time_zone
      - created_at
      - updated_at
      title: UserDetail
...
    UserUpdate:
      properties:
        first_name:
          anyOf:
          - type: string
            maxLength: 120
          - type: 'null'
          title: First Name
        last_name:
          anyOf:
          - type: string
            maxLength: 120
          - type: 'null'
          title: Last Name
        locale:
          anyOf:
          - type: string
          - type: 'null'
          title: Locale
        time_zone:
          anyOf:
          - type: string
          - type: 'null'
          title: Time Zone
        department_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Department Id
        team_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Team Id
      type: object
      title: UserUpdate
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    WorkflowCreate:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        description:
          anyOf:
          - type: string
            maxLength: 1000
          - type: 'null'
          title: Description
        definition:
          additionalProperties: true
          type: object
          title: Definition
      type: object
      required:
      - name
      title: WorkflowCreate
    WorkflowOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        definition:
          additionalProperties: true
          type: object
          title: Definition
        version:
          type: integer
          title: Version
        created_by:
          type: string
          format: uuid
          title: Created By
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - name
      - description
      - definition
      - version
      - created_by
      - created_at
      - updated_at
      title: WorkflowOut
    WorkflowUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 200
            minLength: 1
          - type: 'null'
          title: Name
        description:
          anyOf:
          - type: string
            maxLength: 1000
          - type: 'null'
          title: Description
        definition:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Definition
      type: object
      title: WorkflowUpdate